requirements.txt
# Core Backend
Flask==3.0.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1

# Testing
pytest==8.1.1
pytest-mock==3.12.0
requests==2.31.0

# Linting 
pylint==3.1.0

# E2E Testing (Playwright) 
pytest-playwright==0.7.2
playwright==1.41.2

# Utility Tools
typing_extensions==4.9.0
